Added handling temporary files

+++ task-organizing-files/clean_files.py
@@ -17,6 +17,22 @@ def load_config(filepath):
         print(f"Configuration file was not found at {filepath}. Using default configuration.")
         return DEFAULT_CONFIG
 
+def find_temporary_files(main_dir, directories, tmp_extensions):
+    temporary_files = []
+    all_directores = [main_dir, *directories]
+
+    for dir in all_directores:
+        for dir_path, _, files in os.walk(dir):
+            for file in files:
+                file_path = os.path.join(dir_path, file)
+
+                if any(file.endswith(tmp_extension) for tmp_extension in tmp_extensions):
+                    temporary_files.append(file_path)
+
+    return temporary_files
+
+
+
 def find_empty_files(main_dir, directories):
     empty_files = []
     all_directories = [main_dir, *directories]
@@ -31,11 +47,13 @@ def find_empty_files(main_dir, directories):
 
     return empty_files
 
-def handle_empty_files(empty_files):
+def ask_before_deleting(empty_files, what_to_delete: str):
 
     choice = None
     for empty_file in empty_files:
-        print("Empty file was found at: {empty_file}")
+        if choice == 'an':
+            continue
+        print(f"{what_to_delete.capitalize()} was found at: {empty_file}. Do you want to remove it? ")
 
         if choice == 'ay':
             os.remove(empty_file)
@@ -43,8 +61,8 @@ def handle_empty_files(empty_files):
 
         choice = get_user_input()
         if choice == 'y':
-            print("Removing file {empty_file}")
-            os.remove()
+            print(f"Removing file {empty_file}")
+            os.remove(empty_file)
 
 def get_user_input():
     accepted_values = ['y', 'n', 'ay', 'an']
@@ -60,6 +78,8 @@ def get_user_input():
         print("Invalid input. Please enter your choice again. ")
 
 
+
+
 def parse_arguments():
     parser = ArgumentParser(description="Clean files from given directories")
 
@@ -68,6 +88,7 @@ def parse_arguments():
     parser.add_argument("-c", "--config", help="Path to json configuration file")
 
     parser.add_argument("--empty", action="store_true", help="Search for empty files and suggest deleting them")
+    parser.add_argument("--temporary", action="store_true", help="Search for temporary files and suggest deleting them")
 
     return parser.parse_args()
 
@@ -81,8 +102,12 @@ if __name__ == "__main__":
     print("Configuration: ", config)
 
     if (args.empty):
-        print('dupa')
-        print(find_empty_files(args.main_dir, args.directories))
+        empty_files = find_empty_files(args.main_dir, args.directories)
+        ask_before_deleting(empty_files, "empty file")
+
+    if (args.temporary):
+        temporary_files = find_temporary_files(args.main_dir, args.directories, config['temporary_file_extensions'])
+        ask_before_deleting(temporary_files, "temporary file")
